+++ Assets/Scripts/Classes/Health.cs
@@ -7,22 +7,22 @@ namespace Veganimus.Platformer
         [SerializeField, Tooltip("Indicate if object is the player.")]
         protected bool _isPlayer = false;
         [SerializeField, Tooltip("Hit Points")]
-        protected sbyte _hp;
+        protected int _hp;
         [SerializeField, Tooltip("Lives and/or Energy Tanks")]
-        protected sbyte _lives;
+        protected int _lives;
         [SerializeField, Tooltip("Max Lives and/or Energy Tanks")]
-        protected byte _maxLives;
+        protected int _maxLives;
         [SerializeField, Tooltip("Set Friendliness.")]
         protected CharacterType _characterType;
         protected bool _gameStart = true;
-        protected byte _maxLifeHP = 99;
+        protected int _maxLifeHP = 99;
         protected UIManager _uIManager;
-        public sbyte HP { get { return _hp; } set { _hp = value; } }
+        public int HP { get { return _hp; } set { _hp = value; } }
         public bool IsPlayer { get { return _isPlayer;} }
         
-        public sbyte Lives { get { return _lives; } private set { _lives = value; } }
+        public int Lives { get { return _lives; } private set { _lives = value; } }
 
-        public virtual void Damage(sbyte hpDamage)
+        public virtual void Damage(int hpDamage)
         {
             _hp -= hpDamage;
 
@@ -38,12 +38,12 @@ namespace Veganimus.Platformer
         /// Pass in the <value>amount to heal the implementing object.</value>
         /// </summary>
         /// <param name="amount"></param>
-        public virtual void Heal(sbyte amount)
+        public virtual void Heal(int amount)
         {
             _hp += amount;
             if(_hp + amount > _maxLifeHP && _lives == _maxLives)
             {
-                _hp = (sbyte)_maxLifeHP;
+                _hp = (int)_maxLifeHP;
             }
         }
     }
