+++ Assets/Scripts/Classes/LoadingScreen.cs
@@ -10,7 +10,7 @@ namespace Veganimus.Platformer
         [SerializeField] private Image _progessBar;
         [SerializeField] private SceneLoader _loader;
         [SerializeField] private TMP_Text _loadingProgress;
-        public byte sceneIndex;
+        public int sceneIndex;
 
         private void Start() => StartCoroutine(LoadLevelASync());
